configure.ac Makefile.am changes for AGG definition
Not sure what the original mechanism for running `make` / having `make`
run `gcc`  with `AGG` defined was, but I seem unable to get it to work.
I'm sure that is a lack of knowledge on my part.

The following adds a configure option that will cause calls to `make` to
define `AGG` appropriately.  Usage is:

```
./configure --enable-group-aggregations
make
```